Add git branch + staged + unstaged changes
I don't _love_ this code, I'm doing all of the git work entirely
through zsh built-ins and a) it's slow and b) I really want to
conditionally change the *color* of the git part of my prompt
if the repo is dirty, but I'd have to do that while setting it in
vcs_info (since it's dropped into the prompt as a single var) and
I can't figure out a clean conditional. Will probably swap to a
more powerful script, like git's git-prompt.sh, and ditch most of
this later.